// src/app/models/survey/survey.model.ts
export interface surveyDto {
  q1: number;
  q1freetext: string;

  q2: number;
  q2freetext: string;

  q3: number;
  q3freetext: string;

  q4: number;
  q4freetext: string;

  q5: number;
  q5freetext: string;

  q6: number;
  q6freetext: string;

  q7: number;
  q7freetext: string;

  q8: number;
  q8freetext: string;

  q9: number;
  q9freetext: string;

  q10: number;
  q10freetext: string;

  anythingelse: string;
}